requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-rest-auth==2.1.9
Django==4.2
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
oauthlib==3.2.2
pillow==10.3.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
sqlparse==0.5.0
